requirements.txt
python-dotenv
beautifulsoup4
selenium
